requirements.txt
telethon
python-dotenv
httpx
requests
loguru
